More packet info, recipe setting for metal forge
Added more info into the packet like networkID and values in the metal
forge.

+++ src/common/infinitealloys/TileEntityComputer.java
@@ -52,7 +52,7 @@ public class TileEntityComputer extends TileEntityMachine {
 
 	public TileEntityComputer(int facing) {
 		this();
-		orientation = facing;
+		orientation = (byte)facing;
 	}
 
 	public TileEntityComputer() {
